requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-filter==24.3
django-solo==2.4.0
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
